exit with error status when run with -t (--test)
When xcpretty is run with -t it's a fair assumption
that the input to xcpretty should contain a valid test run.

If the test run is not complete, there are no
tests, or there are failing tests, this commit
causes xcpretty to exit with a status code of 1.

This is useful for use in Travis CI, as it errs
on the side of false negatives, rather than false
positives which are much more of a headache since
they can result in invalid code getting merged.

The current solution of appending exit ${PIPESTATUS[0]}
to the pipe chain fails when xcodebuild exits with
an status code of 0 when the build fails. An easy
way to see that happen is to delete a file from
your working copy without deleting it from the
xcode project.