Fix xcpretty swallowing up errors
We are a grown CI solutions for iOS apps and are gladly using xcpretty.
That said, we have found instances where xcpretty is swallowing up errors making our end users lives rather unpleasant.

For phase script execution:
a. The regular expression was wrong
b. There was no output to be able to diagnose the issue if an error is present

For Compile Storyboards:
a. There was no output to determine while compilation failed.

For Generic Errors
a. We should be spitting out generic errors