Add the JUnit time attribute on a failed tests.
The Sonar JUnit parser fails if the time attribute is not present on
failed tests. This patch adds the time attribute for these tests with a
fixed duration of "0".

Adding the real test duration was considered however it cannot be
done easily with the current xcpretty implementation:  xcpretty parses
Xcode output line by line whereas a failed test outputs at least 2 lines
of log: one that shows the failed test and the cause, another that
displays the duration (quite the same line as a passing test).
A considered approach would be to keep a context and output a
transformed log line only if the context is full (in the case of failed
tests, only if the cause and the duration have been parsed).