Do not erase but update compilation database.
When using the -r json-compilation-database option, if a database already
exists, it is completely rewritten. This has the side effect that when using
incremental compilation, the compilation commands for the files that are not
recompiled are lost. This has the effect that tools using the database like
clang-tidy won't find the compilation commands for file that were not
recompiled.

This commit changes the database generation code so that it updates the database
with the detected commands instead of rewriting it.