Handle ProcessProductPackaging message (#324)
* Handle ProcessProductPackaging message

This message is printed when reading xcent files. It has an arbitrary
length depending on your entitlements. This currently has one issue
where it prints a newline after itself, because it exits on a specific
string after matching as many lines as possible between the start, and
there's no way to capture output after exiting.

* shut up rubocop